% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aSPUmin.R
\name{aSPUmin}
\alias{aSPUmin}
\title{Combined p-value by minP and Fisher approached of aSPU test incorporating multiple sources of functional annotations (aSPUmin)}
\usage{
aSPUmin(Y, X, annot, annot.wt = NULL, cov = NULL, pow = c(1:6),
  B = 1000)
}
\arguments{
\item{Y}{Response or phenotype data. It can be a quantitative trait. A vector with length n (number of observations).}

\item{X}{Genotype or other data; each row for a subject, and each column
for an SNP (or a predictor). The value of each SNP is the # of the copies
for an allele. A matrix with dimension n by k (n : number of observation, k : number of SNPs (or predictors) ). No NAs are allowed. Please impute the NA vslues first.}

\item{annot}{Functional annotations. A matrix with dimension k by r (k :number of SNPs, r : number of annotations).}

\item{annot.wt}{The genome-wide informations. A vector with length r (number of annotations).}

\item{pow}{SNP specific power(gamma values) used in FunSPU test.}

\item{B}{number of permutations.}

\item{Z}{Covariates. A matrix with dimension n by p (n :number of observation, p : number of covariates).}

\item{pow.annot}{Annotation specific power(gamma a values) used in FunSPU test.}
}
\value{
P-values for minP and Fisher approaches combining aSPU tests.
}
\description{
It gives p-values of the minP and Fisher approaches of multiple aSPU tests.
}
\examples{

data(dat)
## example analysis using aSPUmin test on data.ex data.
test.aSPUmin <- aSPUmin(Y=data.ex$Y, X=data.ex$X, annot=data.ex$annot,
 annot.wt=data.ex$annot.wt, pow=c(1:6), B=1000)

test.aSPUmin$minP
# This is p-value of minP approach.

test.aSPUmin$Fisher
# This is p-value of Fisher approach.
}
\references{
Yiding Ma and Peng Wei (2017)
A Powerful and Adaptive Test for Incorporating Multiple Sources of Biological Knowledge into Association Analysis of Whole Genome Sequencing Data (Publication)
}
\seealso{
\code{\link{aSPUwmin}}
}
\author{
Yiding Ma and Peng Wei
}
